// helpers/types.ts
export enum eEthereumNetwork {
  buidlerevm = "buidlerevm",
  kovan = "kovan",
  ropsten = "ropsten",
  main = "main",
  coverage = "coverage",
  hardhat = "hardhat",
  tenderlyMain = "tenderlyMain",
}

export enum eHecoNetwork {
  htestnet = "htestnet",
  heco = "heco",
}

export enum eHooNetwork {
  hoo = "hoo",
}

export enum eArbitrumNetwork {
  amain = "amain",
  arinkeby = "arinkeby",
}

export enum eBscNetwork {
  bsc = "bsc",
  bsctestnet = "bsctestnet",
}

export enum eOkexNetwork {
  okex = "okex",
}

export enum ePolygonNetwork {
  matic = "matic",
  mumbai = "mumbai",
}

export enum eXDaiNetwork {
  xdai = "xdai",
}
// ...
export type iParamsPerNetwork<T> =
  | iEthereumParamsPerNetwork<T>
  | iPolygonParamsPerNetwork<T>
  | iXDaiParamsPerNetwork<T>
  | iHecoParamsPerNetwork<T>
  | iHooParamsPerNetwork<T>
  | iArbitrumParamsPerNetwork<T>
  | iBscParamsPerNetwork<T>
  | iOkexParamsPerNetwork<T>;

export interface iParamsPerNetworkAll<T>
  extends iEthereumParamsPerNetwork<T>,
    iPolygonParamsPerNetwork<T>,
    iXDaiParamsPerNetwork<T>,
    iHooParamsPerNetwork<T>,
    iArbitrumParamsPerNetwork<T>,
    iBscParamsPerNetwork<T>,
    iOkexParamsPerNetwork<T>,
    iHecoParamsPerNetwork<T> {}

export interface iEthereumParamsPerNetwork<T> {
  [eEthereumNetwork.coverage]: T;
  [eEthereumNetwork.buidlerevm]: T;
  [eEthereumNetwork.kovan]: T;
  [eEthereumNetwork.ropsten]: T;
  [eEthereumNetwork.main]: T;
  [eEthereumNetwork.hardhat]: T;
  [eEthereumNetwork.tenderlyMain]: T;
}

export interface iPolygonParamsPerNetwork<T> {
  [ePolygonNetwork.matic]: T;
  [ePolygonNetwork.mumbai]: T;
}

export interface iXDaiParamsPerNetwork<T> {
  [eXDaiNetwork.xdai]: T;
}

export interface iHecoParamsPerNetwork<T> {
  [eHecoNetwork.htestnet]: T;
  [eHecoNetwork.heco]: T;
}

export interface iHooParamsPerNetwork<T> {
  [eHooNetwork.hoo]: T;
}

export interface iArbitrumParamsPerNetwork<T> {
  [eArbitrumNetwork.amain]: T;
  [eArbitrumNetwork.arinkeby]: T;
}

export interface iBscParamsPerNetwork<T> {
  [eBscNetwork.bsc]: T;
  [eBscNetwork.bsctestnet]: T;
}

export interface iOkexParamsPerNetwork<T> {
  [eOkexNetwork.okex]: T;
}